requirements.txt
Django>=5.2.7,<6.0.0
djangorestframework>=3.16.1,<4.0.0
djangorestframework-simplejwt>=5.5.1,<6.0.0
drf-spectacular>=0.28.0,<0.29.0
drf-spectacular-sidecar>=2025.10.1,<2026.0.0
django-filter>=25.2,<26.0
python-dotenv>=1.2.1,<2.0.0
celery>=5.5.3,<6.0.0
redis>=5.0,<6.0
psycopg2-binary>=2.9
django-celery-beat>=2.6